scripts/requirements.txt
boto3>=1.28.0
ruamel.yaml>=0.17.0
PyYAML>=6.0
